Appender rework
This change implements the access to the Appender interface from Java with
the following features:

 - C API is used to access the native Appender
 - necessary C API calls are exposed to Java using JNI wrappers as thin
   as possible - Java calls mirror corresponding C API calls 1 to 1
 - the data chunk interface of the Appender API is used: vector data is
   exposed as a direct ByteBuffer, all primitive appended values are
   written to this buffer from Java without going through JNI + C API (
   that is still used for some types with calls like:
   `duckdb_vector_assign_string_element_len`)
 - Java-side Appender/DataChunk/Vector data structures follow the Go
   lang's Appender implementation closely (with nested arrays
   initialization etc)
 - Java Appender usage is made thread-safe for concurrent Appender or
   Connection closure; `append()` calls are remained not thread-safe (to
   minimize the overhead), but their usage cannot cause the JNI-side
   crash
 - Java API of the new Appender is modeled after the
   `java.lang.StringBuilder` class and intended to be used with method
   chaining
 - support primitive arrays (one and two dimensional)
 - support for `NULL` and `DEFAULT` values
 - type checks between Java types and DB column types are enforced

Note: previous version of the Appender (that internally creates
`Value` instances and appends them one by one) is still available for
compatibility. It can be created using
`Connection#createSingleValueAppender` method. It is marked as
'deprecated' and intended to be removed in future versions.

Testing: existing Appender test suite is extended and adapted to new
Appender API.

Fixes: #84, #100, #110, #139, #157, #163, #219, #249

+++ src/test/java/org/duckdb/TestPrepare.java
@@ -23,9 +23,9 @@ public class TestPrepare {
             ps.setByte(2, (byte) 42);
             ps.setShort(3, (short) 43);
             ps.setInt(4, 44);
-            ps.setLong(5, (long) 45);
+            ps.setLong(5, 45);
             ps.setFloat(6, (float) 4.6);
-            ps.setDouble(7, (double) 4.7);
+            ps.setDouble(7, 4.7);
             ps.setString(8, "four eight");
 
             try (ResultSet rs = ps.executeQuery()) {
@@ -46,7 +46,7 @@ public class TestPrepare {
             ps.setInt(4, 84);
             ps.setLong(5, (long) 85);
             ps.setFloat(6, (float) 8.6);
-            ps.setDouble(7, (double) 8.7);
+            ps.setDouble(7, 8.7);
             ps.setString(8, "eight eight\n\t");
 
             try (ResultSet rs = ps.executeQuery()) {
@@ -67,7 +67,7 @@ public class TestPrepare {
             ps.setObject(4, 84);
             ps.setObject(5, (long) 85);
             ps.setObject(6, (float) 8.6);
-            ps.setObject(7, (double) 8.7);
+            ps.setObject(7, 8.7);
             ps.setObject(8, "𫝼🔥😜䭔🟢");
 
             try (ResultSet rs = ps.executeQuery()) {
